booktracker/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
django-rest-framework==0.1.0
django-test==0.4030
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
